Create type of triangle.txt
Determine the type of triangle (acute-angled, obtuse, rectangular) with the given sides.

Input data
Three natural numbers are given - the sides of the triangle. Each number does not exceed 10,000.

Output
You must output one of the words: “right” for a right-angled triangle, “acute” for an acute-angled triangle, “obtuse” for an obtuse-angled triangle, or “impossible” if the input numbers do not form a triangle.
Sample Input:

3
4
5
Sample Output:

right