Create The maximum amount of digits. (nested loop).txt
The maximum amount of digits.
Among natural numbers from 1 to N (including N), find the largest number in terms of the sum of the digits. Print this number and the sum of its digits.

Input data
The first line contains the only positive integer N not exceeding 10 ^ 6

Output
Print the number itself, which has the maximum sum of digits and the sum of digits on one line, separated by a space. If there are several such numbers, print the minimum of them.

Sample Input:

20
Sample Output:

19 10